refactor(build): remove default argument
This commit removes the `mode` argument specified. The `mode` argument
with the type OpenTextMode defaults to read mode, and thus shouldn't be
specified.

Similarly, to how this:

open("foo.c", errors=None, newline=None, closefd=True)

is the same as this:

open("foo.c")

because `closefd`, `errors`, `newline` are the same as their defaults.